Create README.md
The project builds upon a number of technologies to provide for a way to identify if a vegetable has matured to a point to be picked from its host plant using remote sensing and artificial intelligence.  Leveraging and brokering technologies from a number of sources including but not exclusively Farmbot.io, Structure.io and NVidia.com.  The challenge will be to combine and code a solution from a number of party sensors to build a solution that will grow plants from seedlings , measure and sense plant and vegetable health and determine if vegetable is ready to be harvested.